refactor: small refactor improvements: move data to folder to src folder, styleguide renaming
