+++ src/app/components/externalLink/external-link.scss
@@ -12,7 +12,7 @@
     color: #000;
   }
 
-  & .icon-wrapper {
+  .icon-wrapper {
     display: inline-block;
     width: 25px;
     margin-left: 5px;
