+++ src/data/communication.data.ts
@@ -1,6 +1,6 @@
 export const RS_DOCS_COMMUNICATION_LINK = 'https://docs.rs.school/#/rs-school-chats';
 
-export const discordLinks = {
+export const DISCORD_LINKS = {
   javascript: 'https://discord.com/invite/QvEYg7EaQ4',
   'javascript-en': 'https://discord.com/invite/uW5cCHR',
   react: 'https://discord.com/invite/zyRcphs3px',
@@ -11,12 +11,4 @@ export const discordLinks = {
   'aws cloud dev': 'https://discord.com/invite/WEZxwRa4J6',
 };
 
-export type CourseNames =
-  | 'javascript'
-  | 'javascript-en'
-  | 'react'
-  | 'react ru'
-  | 'angular'
-  | 'node.js'
-  | 'aws fundamentals'
-  | 'aws cloud dev';
+export type CourseNames = keyof typeof DISCORD_LINKS;
